Edit Method DeleteCategoryByIdAsync

+++ Team-Library/BackEnd/TeamLibrary.API/Service/Implementation/CategoryService.cs
@@ -45,7 +45,11 @@ namespace TeamLibrary.API.Service.Implementation
 
         public async Task DeleteCategoryByIdAsync(int categoryId)
         {
-
+            var category = _repository.GetCategoryByIdAsync(categoryId);
+            if (category == null) 
+            {
+                return;
+            }
             await _repository.DeleteCategoryByIdAsync(categoryId);
         }
